Address review: extract block-scalar helpers, test quoted inline lambda
- Move isEditableLambdaBlock / lambdaValueFromBlock / blockScalarValue
  into yaml-block-scalar-value.ts so yaml-section-reader.ts stays under
  the 500-600 line cap (was pushed to 636; back to 600).
- Add a regression test for the quoted inline `!lambda '<body>'` form.
