+++ src/util/yaml-section-reader.ts
@@ -6,9 +6,13 @@
  * facade in yaml-section-values.ts.
  */
 
-import type { LambdaValue } from "../api/types/automations.js";
 import { ESPHOME_YAML_INDENT } from "./esphome-yaml-lang.js";
 import { LIST_SECTIONS } from "./section-entry-overrides.js";
+import {
+  blockScalarValue,
+  isEditableLambdaBlock,
+  lambdaValueFromBlock,
+} from "./yaml-block-scalar-value.js";
 import { parseFlowList, parseScalar, splitInlineComment } from "./yaml-scalar.js";
 import {
   _detectListItemChildIndent,
@@ -37,46 +41,6 @@ import {
 } from "./yaml-section-list.js";
 import { type KeySpan, type ParsedSection } from "./yaml-section-splice.js";
 import { YamlRawValue } from "./yaml-serialize.js";
-
-/**
- * True for the canonical strip-chomped literal block (``|-``) — the
- * only marker the form's lambda editor (and the serializer) emit.
- * Other markers (folded ``>``, keep ``|+``) carry distinct YAML
- * semantics the editor would silently normalise to ``|-``, so they
- * stay opaque ``YamlRawValue`` blocks instead of editable lambdas.
- */
-const isEditableLambdaBlock = (header: {
-  tag: string | undefined;
-  marker: string;
-}): boolean => header.tag === "!lambda" && header.marker === "|-";
-
-/**
- * Build a ``LambdaValue`` sentinel from a captured ``!lambda |-`` block
- * body. Reuses ``YamlRawValue.body`` to strip the block's common
- * indent (so ``          return 0.01;`` becomes ``return 0.01;``);
- * only trailing newlines are dropped (the ``|-`` strip chomp), not
- * trailing spaces/tabs on the last line.
- */
-const lambdaValueFromBlock = (bodyLines: string[]): LambdaValue => ({
-  _lambda: new YamlRawValue(bodyLines).body.replace(/\n+$/, ""),
-  _tag: "!lambda",
-});
-
-/**
- * Turn a (possibly tagged) block-scalar header + its captured body
- * lines into a form value: a ``LambdaValue`` for a canonical
- * ``!lambda |-`` so the value stays editable, a ``YamlRawValue``
- * (carrying the verbatim header) for any other tag, marker, or a
- * bare ``|-``/``>`` block.
- */
-const blockScalarValue = (
-  header: { tag: string | undefined; marker: string },
-  rawHeader: string,
-  bodyLines: string[]
-): LambdaValue | YamlRawValue =>
-  isEditableLambdaBlock(header)
-    ? lambdaValueFromBlock(bodyLines)
-    : new YamlRawValue(bodyLines, rawHeader);
 
 /**
  * Dispatch a YAML list block (``key:\n  - …``) into the right
